requirements.txt
absl-py==2.2.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.13
decorator==5.2.1
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h5py==3.13.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
promise==2.3
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyzmq==26.4.0
requests==2.32.3
rich==13.9.4
setuptools==78.0.2
simple-parsing==0.1.7
six==1.17.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.8
tensorflow-macos==2.16.2
tensorflow-metadata==1.17.0
tensorflow-metal==1.2.0
termcolor==2.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
